Issue #8: Added demo request (`users`) on `MainSurveysList` (will be replaced by `surveys` request later), refactored `user` data model in mirage server.
