+++ src/entities/AppMenu/types/AppMenuTypes.ts
@@ -1,16 +1,4 @@
-import {
-  Menu,
-  HelpOutline,
-  // DriveFolderUpload,
-  DarkMode,
-  LightMode,
-  BugReport,
-  // BarChart,
-  SvgIconComponent,
-  Home,
-  Login,
-  Logout,
-} from '@mui/icons-material';
+import { SvgIconComponent } from '@mui/icons-material';
 
 export interface TAppMenuItem {
   icon?: SvgIconComponent;
