+++ src/mirage-server/README.md
@@ -1,0 +1,6 @@
+# Mirage api server
+
+## See:
+
+- https://miragejs.com/quickstarts/react/develop-an-app/
+- https://stephencharlesweiss.com/miragejs-typescript-setup
