+++ src/mirage-server/user/README.md
@@ -1,0 +1,3 @@
+# Mirage data model `user`
+
+Used for features testing.
