+++ src/mirage-server/user/index.ts
@@ -1,0 +1,4 @@
+export * from './factories';
+export * from './models';
+export * from './routes';
+export * from './seeds';
