+++ src/mirage-server/user/models.ts
@@ -1,0 +1,6 @@
+import { Model } from 'miragejs';
+
+import { TUser } from 'src/core/types';
+import { ModelDefinition } from 'miragejs/-types';
+
+export const UserModel: ModelDefinition<TUser> = Model.extend({});
