+++ src/mirage-server/user/seeds.ts
@@ -1,0 +1,5 @@
+import { Server } from 'miragejs';
+
+export function userSeeds(server: Server) {
+  server.createList('user', 1);
+}
